requirements.txt
flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
numpy==1.26.2
faiss-cpu
dtaidistance
psycopg2-binary
